+++ gazeplay/src/main/java/net/gazeplay/GameSpec.java
@@ -3,7 +3,7 @@ package net.gazeplay;
 import javafx.scene.Group;
 import javafx.scene.Scene;
 import javafx.scene.control.ChoiceBox;
-import utils.games.stats.Stats;
+import net.gazeplay.utils.stats.Stats;
 
 public class GameSpec {
