+++ gazeplay/src/main/java/net/gazeplay/games/creampie/CreampieStats.java
@@ -1,7 +1,7 @@
 package net.gazeplay.games.creampie;
 
 import javafx.scene.Scene;
-import utils.games.stats.ShootGamesStats;
+import net.gazeplay.utils.stats.ShootGamesStats;
 
 public class CreampieStats extends ShootGamesStats {
