+++ gazeplay/src/main/java/net/gazeplay/games/magiccards/MagicCardsGamesStats.java
@@ -1,7 +1,7 @@
 package net.gazeplay.games.magiccards;
 
 import javafx.scene.Scene;
-import utils.games.stats.HiddenItemsGamesStats;
+import net.gazeplay.utils.stats.HiddenItemsGamesStats;
 
 public class MagicCardsGamesStats extends HiddenItemsGamesStats {
