+++ gazeplay/src/main/java/net/gazeplay/utils/stats/HiddenItemsGamesStats.java
@@ -1,36 +1,15 @@
-package utils.games.stats;
+package net.gazeplay.utils.stats;
 
 import javafx.scene.Scene;
 import utils.games.Utils;
 
 import java.io.PrintWriter;
 
-public class ShootGamesStats extends Stats{
+public class HiddenItemsGamesStats extends Stats{
 
-    protected int nbUnCountedShoots;
-
-    public ShootGamesStats(Scene scene) {
+    public HiddenItemsGamesStats(Scene scene) {
 
         super(scene);
-
-        nbUnCountedShoots = 0;
-    }
-
-    public void incNbGoals(){
-
-        long last = System.currentTimeMillis() - beginTime;
-        if(last>100) {
-            nbGoals++;
-            length += last;
-            lengthBetweenGoals.add((new Long(last)).intValue());
-        }else{
-
-            nbUnCountedShoots++;
-        }
-    }
-
-    public int getNbUnCountedShoots() {
-        return nbUnCountedShoots;
     }
 
     @Override
@@ -54,12 +33,10 @@ public class ShootGamesStats extends Stats{
         out.print(',');
         out.print("Standard Déviation");
         out.print(',');
-        out.print("Uncounted Shoots");
-        out.print(',');
         for(int i = 0; i < lengthBetweenGoals.size(); i++) {
             out.print("shoot ");
             out.print(i);
-            out.print(',');
+            out.print(",");
         }
         out.println();
 
@@ -76,8 +53,6 @@ public class ShootGamesStats extends Stats{
         out.print(getAverageLength());
         out.print(',');
         out.print(getSD());
-        out.print(',');
-        out.print(getNbUnCountedShoots());
         out.print(',');
         printLengthBetweenGoalsToString(out);
         out.println();
