+++ gazeplay/src/main/java/net/gazeplay/utils/stats/ShootGamesStats.java
@@ -1,15 +1,36 @@
-package utils.games.stats;
+package net.gazeplay.utils.stats;
 
 import javafx.scene.Scene;
 import utils.games.Utils;
 
 import java.io.PrintWriter;
 
-public class HiddenItemsGamesStats extends Stats{
+public class ShootGamesStats extends Stats{
 
-    public HiddenItemsGamesStats(Scene scene) {
+    protected int nbUnCountedShoots;
+
+    public ShootGamesStats(Scene scene) {
 
         super(scene);
+
+        nbUnCountedShoots = 0;
+    }
+
+    public void incNbGoals(){
+
+        long last = System.currentTimeMillis() - beginTime;
+        if(last>100) {
+            nbGoals++;
+            length += last;
+            lengthBetweenGoals.add((new Long(last)).intValue());
+        }else{
+
+            nbUnCountedShoots++;
+        }
+    }
+
+    public int getNbUnCountedShoots() {
+        return nbUnCountedShoots;
     }
 
     @Override
@@ -33,10 +54,12 @@ public class HiddenItemsGamesStats extends Stats{
         out.print(',');
         out.print("Standard Déviation");
         out.print(',');
+        out.print("Uncounted Shoots");
+        out.print(',');
         for(int i = 0; i < lengthBetweenGoals.size(); i++) {
             out.print("shoot ");
             out.print(i);
-            out.print(",");
+            out.print(',');
         }
         out.println();
 
@@ -53,6 +76,8 @@ public class HiddenItemsGamesStats extends Stats{
         out.print(getAverageLength());
         out.print(',');
         out.print(getSD());
+        out.print(',');
+        out.print(getNbUnCountedShoots());
         out.print(',');
         printLengthBetweenGoalsToString(out);
         out.println();
